+++ apps/cold-api/src/platform/modules/resources/surveys/surveys.module.ts
@@ -4,10 +4,13 @@ import { JwtStrategy } from '@coldpbc/nest';
 import { SurveysController } from './surveys.controller';
 import { SurveysService } from './surveys.service';
 import { SurveysRabbitService } from './surveys.rabbit';
+import { ScoringModule } from './scoring/scoring.module';
+import { SurveyFilterService } from './filter/survey.filter.service';
 
 @Module({
+  imports: [ScoringModule],
   controllers: [SurveysController],
-  providers: [SurveysService, JwtService, JwtStrategy, SurveysRabbitService],
+  providers: [SurveysService, JwtService, JwtStrategy, SurveysRabbitService, SurveyFilterService],
   exports: [SurveysService, JwtService, JwtStrategy, SurveysRabbitService],
 })
 export class SurveysModule {}